+++ crates/core/src/lib.rs
@@ -16,7 +16,9 @@ pub type Result<T> = std::result::Result<T, error::Error>;
 
 pub use alloy;
 pub use contender_bundle_provider::bundle::BundleType;
-pub use orchestrator::{Contender, ContenderCtx, RunOpts};
+pub use orchestrator::{
+    Contender, ContenderCtx, Initialized, LifecyclePhase, PhaseMarker, RunOpts, Uninitialized,
+};
 pub use tokio::task as tokio_task;
 pub use tokio_util::sync::CancellationToken;